added website_editor to new site system
